Move CollectedDataAggregatingJob to Jobs folder
